paper_decomposer/requirements.txt
# Paper Decomposer & Notebook Builder Requirements
# Install with: pip install -r requirements.txt

# Core dependencies (required)
nbformat>=5.7.0
pymupdf>=1.23.0

# Notebook execution (recommended)
papermill>=2.4.0
nbconvert>=7.0.0
ipykernel>=6.20.0

# Alternative PDF extraction (optional)
pdfminer.six>=20221105

# Testing (development)
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Note: AgentRLM core dependencies (rlm, anthropic, openai, etc.)
# should already be installed from the main pyproject.toml
